+++ spec/acceptance/1_installs_spec.rb
@@ -15,11 +15,11 @@ else
   java_url = download_url
 end
 
-# We add the sleeps everywhere to give stash enough
+# We add the sleeps everywhere to give bitbucket enough
 # time to install/upgrade/run migration tasks/start
 
-describe 'stash', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do
-  it 'installs with defaults and context /stash1' do
+describe 'bitbucket', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do
+  it 'installs with defaults and context /bitbucket1' do
     pp = <<-EOS
       $jh = $osfamily ? {
         default   => '/opt/java',
@@ -35,9 +35,9 @@ describe 'stash', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do
         version             => '9.4',
       }->
       class { 'postgresql::server': } ->
-      postgresql::server::db { 'stash':
-        user     => 'stash',
-        password => postgresql_password('stash', 'password'),
+      postgresql::server::db { 'bitbucket':
+        user     => 'bitbucket',
+        password => postgresql_password('bitbucket', 'password'),
       } ->
       deploy::file { 'jdk-8u45-linux-x64.tar.gz':
         target          => '/opt/java',
@@ -46,22 +46,22 @@ describe 'stash', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do
         download_timout => 1800,
         strip           => true,
       } ->
-      class { 'stash':
+      class { 'bitbucket':
         download_url   => #{download_url},
         checksum      => 'a80cbed70843e9a394867adc800f7704',
         version       => '3.9.2',
         javahome      => $jh,
-        context_path  => '/stash1',
+        context_path  => '/bitbucket1',
         backupclient_url => #{download_url},
       }
-      include ::stash::facts
+      include ::bitbucket::facts
   EOS
     apply_manifest(pp, :catch_failures => true)
     sleep 180
-    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/stash1'
+    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/bitbucket1'
     apply_manifest(pp, :catch_failures => true)
     sleep 180
-    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/stash1', :acceptable_exit_codes => [0]
+    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/bitbucket1', :acceptable_exit_codes => [0]
     apply_manifest(pp, :catch_changes => true)
   end
 
@@ -77,30 +77,30 @@ describe 'stash', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do
     it { should be_installed }
   end
 
-  describe service('stash') do
+  describe service('bitbucket') do
     it { should be_enabled }
   end
 
-  describe user('stash') do
+  describe user('bitbucket') do
     it { should exist }
-    it { should belong_to_group 'stash' }
+    it { should belong_to_group 'bitbucket' }
     it { should have_login_shell '/bin/bash' }
   end
 
-  describe command('curl http://localhost:7990/stash1/setup') do
-    its(:stdout) { should match(/This is the base URL of this installation of Stash/) }
+  describe command('curl http://localhost:7990/bitbucket1/setup') do
+    its(:stdout) { should match(/This is the base URL of this installation of Bitbucket/) }
   end
 
-  describe command('facter -p stash_version') do
+  describe command('facter -p bitbucket_version') do
     its(:stdout) { should match(/3\.9\.2/) }
   end
 
   describe cron do
-    it { should have_entry('0 5 * * * /opt/java/bin/java -Dstash.password="password" -Dstash.user="admin" -Dstash.baseUrl="http://localhost:7990" -Dstash.home=/home/stash -Dbackup.home=/opt/stash-backup/archives -jar /opt/stash-backup/stash-backup-client-1.9.1/stash-backup-client.jar').with_user('stash') }
+    it { should have_entry('0 5 * * * /opt/java/bin/java -Dbitbucket.password="password" -Dbitbucket.user="admin" -Dbitbucket.baseUrl="http://localhost:7990" -Dbitbucket.home=/home/bitbucket -Dbackup.home=/opt/bitbucket-backup/archives -jar /opt/bitbucket-backup/bitbucket-backup-client-1.9.1/bitbucket-backup-client.jar').with_user('bitbucket') }
   end
 
-  describe file('/opt/stash-backup/stash-backup-client-1.9.1/stash-backup-client.jar') do
+  describe file('/opt/bitbucket-backup/bitbucket-backup-client-1.9.1/bitbucket-backup-client.jar') do
     it { should be_file }
-    it { should be_owned_by 'stash' }
+    it { should be_owned_by 'bitbucket' }
   end
 end