+++ spec/acceptance/2_upgrades_spec.rb
@@ -7,11 +7,11 @@ else
   download_url = 'undef'
 end
 
-# We add the sleeps everywhere to give stash enough
+# We add the sleeps everywhere to give bitbucket enough
 # time to install/upgrade/run migration tasks/start
 
-describe 'stash', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do
-  it 'upgrades to 4.0.2 with defaults and context /stash1' do
+describe 'bitbucket', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do
+  it 'upgrades to 4.0.2 with defaults and context /bitbucket1' do
     pp_update = <<-EOS
       if versioncmp($::puppetversion,'3.6.1') >= 0 {
         $allow_virtual_packages = hiera('allow_virtual_packages',false)
@@ -22,23 +22,23 @@ describe 'stash', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do
       $jh = $osfamily ? {
         default   => '/opt/java',
       }
-      class { 'stash':
+      class { 'bitbucket':
         version       => '3.11.4',
         deploy_module => 'staging',
         download_url   => #{download_url},
         javahome      => $jh,
-        context_path  => '/stash1',
+        context_path  => '/bitbucket1',
       }
-      include ::stash::facts
+      include ::bitbucket::facts
     EOS
     apply_manifest(pp_update, :catch_failures => true)
-    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/stash1'
+    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/bitbucket1'
     sleep 180
     apply_manifest(pp_update, :catch_failures => true)
-    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/stash1', :acceptable_exit_codes => [0]
+    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/bitbucket1', :acceptable_exit_codes => [0]
     sleep 120
     apply_manifest(pp_update, :catch_changes => true)
-    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/stash1', :acceptable_exit_codes => [0]
+    shell 'wget -q --tries=20 --retry-connrefused --read-timeout=10 localhost:7990/bitbucket1', :acceptable_exit_codes => [0]
   end
 
   describe process('java') do
@@ -53,27 +53,27 @@ describe 'stash', :unless => UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do
     it { should be_installed }
   end
 
-  describe service('stash') do
+  describe service('bitbucket') do
     it { should be_enabled }
   end
 
-  describe user('stash') do
+  describe user('bitbucket') do
     it { should exist }
   end
 
-  describe user('stash') do
-    it { should belong_to_group 'stash' }
+  describe user('bitbucket') do
+    it { should belong_to_group 'bitbucket' }
   end
 
-  describe user('stash') do
+  describe user('bitbucket') do
     it { should have_login_shell '/bin/bash' }
   end
 
-  describe command('curl http://localhost:7990/stash1/setup') do
-    its(:stdout) { should match(/This is the base URL of this installation of Stash/) }
+  describe command('curl http://localhost:7990/bitbucket1/setup') do
+    its(:stdout) { should match(/This is the base URL of this installation of Bitbucket/) }
   end
 
-  describe command('facter -p stash_version') do
+  describe command('facter -p bitbucket_version') do
     its(:stdout) { should match(/3\.11\.4/) }
   end
 end