+++ spec/classes/bitbucket_facts_spec.rb
@@ -1,0 +1,49 @@
+require 'spec_helper'
+
+describe 'bitbucket::facts', :type => :class do
+  context 'supported operating systems' do
+    on_supported_os.each do |os, facts|
+      context "on #{os} #{facts}" do
+        let(:facts) do
+          facts
+        end
+        regexp_pe = %r{^#!/opt/puppet/bin/ruby$}
+        regexp_oss = %r{^#!/usr/bin/env ruby$}
+        pe_external_fact_file = '/etc/puppetlabs/facter/facts.d/bitbucket_facts.rb'
+        external_fact_file = '/etc/facter/facts.d/bitbucket_facts.rb'
+
+        it { should contain_file(external_fact_file) }
+
+        # Test puppet enterprise shebang generated correctly
+        context 'with puppet enterprise' do
+          let(:facts) do
+            facts.merge(:puppetversion => '3.4.3 (Puppet Enterprise 3.2.1)')
+          end
+          it do
+            should contain_file(pe_external_fact_file) \
+              .with_content(regexp_pe)
+          end
+        end
+
+        ## Test puppet oss shebang generated correctly
+        context 'with puppet oss' do
+          it do
+            should contain_file(external_fact_file) \
+              .with_content(regexp_oss)
+              .with_content(%r{7990/rest/api/})
+          end
+        end
+
+        context 'with context' do
+          let(:params) do
+            { :context_path => '/bitbucket' }
+          end
+          it do
+            should contain_file(external_fact_file) \
+              .with_content(%r{7990/bitbucket/rest/api/})
+          end
+        end
+      end
+    end
+  end
+end